challenge 14 part 2

+++ src/main/java/dockingData/Instruction.java
@@ -5,5 +5,5 @@ import java.util.Map;
 
 public interface Instruction {
 
-    void execute(Mask mask, Map<Integer, BigInteger> memory);
+    void execute(Mask mask, Map<BigInteger, BigInteger> memory);
 }
